put encryption in its own file
